server/requirements.txt
python-dotenv supabase
langchain
langchain-community
langchain-groq
flask
flask_cors
pymupdf
langchain-cohere
pymupdf
neo4j